requirements.txt
Django==2.1.5
django-bootstrap-form==3.4
django-braces==1.13.0
mysql==0.0.2
mysql-connector-python==8.0.13
mysqlclient==1.3.14
Pillow==5.4.1
protobuf==3.6.1
pytz==2018.9
six==1.12.0
Unidecode==1.0.23
